add icon | trying to complete things | fixing issues | add chart | may contain some bugs

+++ app/src/main/java/com/codecollapse/currencyconverter/data/model/currency/fluctuation/USD.kt
@@ -1,7 +1,7 @@
 package com.codecollapse.currencyconverter.data.model.currency.fluctuation
 
 data class USD(
-    val change: Int,
+    val change: Double,
     val change_pct: Double,
     val end_rate: Double,
     val start_rate: Double
